+++ src/main/kotlin/pl/exbook/exbook/stock/adapter/rest/dto/CreateStockRequest.kt
@@ -1,5 +1,5 @@
 package pl.exbook.exbook.stock.adapter.rest.dto
 
 data class CreateStockRequest(
-    val startQuantity: Int
+    val startQuantity: Long
 )
